Move GMV table related helper methods into a separate class.
